requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.1.47
langchain-openai==0.1.4
langsmith==0.1.52
openai==1.24.1
orjson==3.10.1
packaging==23.2
pydantic==2.7.1
pydantic_core==2.18.2
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
sniffio==1.3.1
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
